fix: fixing the supply formula register and edit

+++ src/main/java/org/sistema/arroz/riceservice/modules/supplyFormula/application/port/in/SupplyFormulaToRegister.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class SupplyFormulaToRegister {
     Long supplyId;
-    Long cantForHectare;
+    Double cantForHectare;
 }
